Segregate library code from test application code.  Folderify library code and update includes accordingly.  Rebuild, verify still builds and runs correctly.
