+++ cpp/src/headers/uint128_test_program.hpp
@@ -6,14 +6,14 @@
 #include <string_view>
 #include <limits>
 #include <cstdint>
-#include "numerics.hpp"
-#include "uint128.hpp"
+#include <cjm/numerics/numerics.hpp>
+#include <cjm/numerics/uint128.hpp>
 #include "int128_tests.hpp"
 #include <boost/io/ios_state.hpp>
-#include "cjm_numeric_concepts.hpp"
+#include <cjm/numerics/cjm_numeric_concepts.hpp>
 #include <exception>
 #include <stdexcept>
-#include "cjm_string.hpp"
+#include <cjm/string/cjm_string.hpp>
 #include <fstream>
 #include <locale>
 #include <tuple>
