--- csharp/bin/info.txt
@@ -1,1 +1,0 @@
-CSharp binaries here.
