--- csharp/info.txt
@@ -1,1 +1,0 @@
-csharp here
