--- csharp/src/info.txt
@@ -1,1 +1,0 @@
-CSharp projects. Etc.
